Add ParserTest, Parser, and Node classes to improve code structuring
In this commit, we added new classes to the codebase to make the programming structure more modular. ParserTest.scala was added to contain necessary test cases for parsing PlantUML diagrams. Parser.scala was incorporated to parse the tokens generated by the lexer, into an Abstract Syntax Tree (AST). Additionally, the Node.scala was included to define the structure of the AST. This encapsulation of code ensures that the code is easier to understand, modify and test.